refactor PathGraph
