+++ SeeSharp/Integrators/Util/PathGraph/LightPathNode.cs
@@ -1,0 +1,7 @@
+namespace SeeSharp.Integrators.Util;
+
+public class LightPathNode(PathVertex lightVertex) : PathGraphNode(lightVertex.Point.Position) {
+    public PathVertex LightVertex { get; } = lightVertex;
+
+    public override RgbColor ComputeVisualizerColor() => RgbColor.SrgbToLinear(228, 135, 17);
+}
